bugfix
Add `Shell` field to struct `Task`, now tasks support the option
of `shell -c command...`

`RuntimeTask` contains now `Output` and `Error` that represents
what you may think.

The method `WaitPoll` now sends over two channels of type `CmdDoneChan`
that contains the task id, the output, the error.

The server receiver taks the output and adds it to the struct. This cannot
be done directly in `WaitPoll` because at the same time the struct could be
accessed by the server in a poll request. The server uses lock, therefore
we'll do it.

Add tests for task.